Move store setup into store module
